EP-3475 reorganise "partitionedjob" module structure
Trying to avoid future circular import issues
